chore: update code

+++ contracts/src/Groth16Verifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 20491192805390485299153009773594534940189261866228447918068658471970481763042;
-    uint256 constant alphay  = 9383485363053290200918347156157836566562967994039712273449902621266178545958;
-    uint256 constant betax1  = 4252822878758300859123897981450591353533073413197771768651442665752259397132;
-    uint256 constant betax2  = 6375614351688725206403948262868962793625744043794305715222011528459656738731;
-    uint256 constant betay1  = 21847035105528745403288232691147584728191162732299865338377159692350059136679;
-    uint256 constant betay2  = 10505242626370262277552901082094356697409835680220590971873171140371331206856;
+    uint256 constant alphax = 20491192805390485299153009773594534940189261866228447918068658471970481763042;
+    uint256 constant alphay = 9383485363053290200918347156157836566562967994039712273449902621266178545958;
+    uint256 constant betax1 = 4252822878758300859123897981450591353533073413197771768651442665752259397132;
+    uint256 constant betax2 = 6375614351688725206403948262868962793625744043794305715222011528459656738731;
+    uint256 constant betay1 = 21847035105528745403288232691147584728191162732299865338377159692350059136679;
+    uint256 constant betay2 = 10505242626370262277552901082094356697409835680220590971873171140371331206856;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,54 +42,57 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 17622675438492827127568598540027857190649682581243915662708310377971049831019;
     uint256 constant deltay2 = 11271953937962341515700148701758561042018064035482955181940969953518782515622;
 
-    
     uint256 constant IC0x = 6953914383719176777209756176816967861961493249371537102563104148933058496701;
     uint256 constant IC0y = 2275319412750954094205726726340799273368013840363181894561481880257774151044;
-    
+
     uint256 constant IC1x = 2062366990153923490283368264435582205428990318704307301793309971318422865261;
     uint256 constant IC1y = 7954407330986699865085305802412238755172941860077566561813193134852070234660;
-    
+
     uint256 constant IC2x = 7277163273925376978734489205647765208170320621184756867764021546915019107506;
     uint256 constant IC2y = 9756858811342895677810878632291397221112936758330324397703461557636748946912;
-    
+
     uint256 constant IC3x = 11253048780538458402462954608101614969612973239429111913846146753170283877145;
     uint256 constant IC3y = 12036242222784129842568229776037893491823451895509222193203296034155180911459;
-    
+
     uint256 constant IC4x = 2842315070580404292578797809656572669344090529607152477226443738503818864610;
     uint256 constant IC4y = 8755311665314932290665981377062820988596380444077865265540204763830193950147;
-    
+
     uint256 constant IC5x = 1138456160897878588286454275086170177880797902727963155236835102417565144395;
     uint256 constant IC5y = 5479747217790400278759036259639553788338379988068487707838862688863451310084;
-    
+
     uint256 constant IC6x = 10128072334320644657703649117026390101159996681557343230076373304860928551443;
     uint256 constant IC6y = 1476388856485940416079824337221633396713314802183667397496729432495730816176;
-    
+
     uint256 constant IC7x = 10129688599150200921266402448957360603239383281238715299030776150279910214313;
     uint256 constant IC7y = 724975737698502137542441806476865952638976008343412762679644173518894477723;
-    
+
     uint256 constant IC8x = 1404364701069320486961712128405511700095575572180794359021979731506299473980;
     uint256 constant IC8y = 13477263532745102253891277350256600084803543276804921964288109800749914549737;
-    
+
     uint256 constant IC9x = 12637536623836026982633940626983094389465361038144442722071890153320885087570;
     uint256 constant IC9y = 11226802851420477857957340465260499807224327722929225689002326086078054374979;
-    
+
     uint256 constant IC10x = 1780073203021385115857595211048419443852461067416125925234938454560989676655;
     uint256 constant IC10y = 10108211957793890164094779743479739811636962281000832249776895396489056229058;
-    
+
     uint256 constant IC11x = 17062830964698477197450188330756010867107452613808217050468998748151349235499;
     uint256 constant IC11y = 2187599036811915512844019140123249312285309866266931020513221683968018357480;
-    
+
     uint256 constant IC12x = 177953299425642802924695164021044706412407335653518702442625117161541598013;
     uint256 constant IC12y = 9667688815034360870002994793437730474451382635817554769193792726099186173892;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[12] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[12] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -97,7 +100,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -132,31 +135,30 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_pVk, IC4x, IC4y, calldataload(add(pubSignals, 96)))
-                
+
                 g1_mulAccC(_pVk, IC5x, IC5y, calldataload(add(pubSignals, 128)))
-                
+
                 g1_mulAccC(_pVk, IC6x, IC6y, calldataload(add(pubSignals, 160)))
-                
+
                 g1_mulAccC(_pVk, IC7x, IC7y, calldataload(add(pubSignals, 192)))
-                
+
                 g1_mulAccC(_pVk, IC8x, IC8y, calldataload(add(pubSignals, 224)))
-                
+
                 g1_mulAccC(_pVk, IC9x, IC9y, calldataload(add(pubSignals, 256)))
-                
+
                 g1_mulAccC(_pVk, IC10x, IC10y, calldataload(add(pubSignals, 288)))
-                
+
                 g1_mulAccC(_pVk, IC11x, IC11y, calldataload(add(pubSignals, 320)))
-                
+
                 g1_mulAccC(_pVk, IC12x, IC12y, calldataload(add(pubSignals, 352)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -182,7 +184,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -199,7 +200,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -209,37 +209,36 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
+
             checkField(calldataload(add(_pubSignals, 128)))
-            
+
             checkField(calldataload(add(_pubSignals, 160)))
-            
+
             checkField(calldataload(add(_pubSignals, 192)))
-            
+
             checkField(calldataload(add(_pubSignals, 224)))
-            
+
             checkField(calldataload(add(_pubSignals, 256)))
-            
+
             checkField(calldataload(add(_pubSignals, 288)))
-            
+
             checkField(calldataload(add(_pubSignals, 320)))
-            
+
             checkField(calldataload(add(_pubSignals, 352)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
