+++ contracts/src/Groth16VerifierConsensus.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16VerifierConsensus {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 20491192805390485299153009773594534940189261866228447918068658471970481763042;
-    uint256 constant alphay  = 9383485363053290200918347156157836566562967994039712273449902621266178545958;
-    uint256 constant betax1  = 4252822878758300859123897981450591353533073413197771768651442665752259397132;
-    uint256 constant betax2  = 6375614351688725206403948262868962793625744043794305715222011528459656738731;
-    uint256 constant betay1  = 21847035105528745403288232691147584728191162732299865338377159692350059136679;
-    uint256 constant betay2  = 10505242626370262277552901082094356697409835680220590971873171140371331206856;
+    uint256 constant alphax = 20491192805390485299153009773594534940189261866228447918068658471970481763042;
+    uint256 constant alphay = 9383485363053290200918347156157836566562967994039712273449902621266178545958;
+    uint256 constant betax1 = 4252822878758300859123897981450591353533073413197771768651442665752259397132;
+    uint256 constant betax2 = 6375614351688725206403948262868962793625744043794305715222011528459656738731;
+    uint256 constant betay1 = 21847035105528745403288232691147584728191162732299865338377159692350059136679;
+    uint256 constant betay2 = 10505242626370262277552901082094356697409835680220590971873171140371331206856;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,57 +42,60 @@ contract Groth16VerifierConsensus {
     uint256 constant deltay1 = 13667536551028573532827105193205093851617931505006871028581969635663450772708;
     uint256 constant deltay2 = 16988259380910557064579901939963386663340170018992993638976987797305388656598;
 
-    
     uint256 constant IC0x = 4785838716472399617204733838708262703003065978316329083665020396176978243197;
     uint256 constant IC0y = 11032342731255774259265987146460494618493968260008566620092944243790138791377;
-    
+
     uint256 constant IC1x = 10944642856064446164116778144819434829929521350161821869435888642277228444872;
     uint256 constant IC1y = 10308683455486076934420052484709980972042783899927381137285558703137862428602;
-    
+
     uint256 constant IC2x = 11568132684523963392859709958427136497730999432942671869854675654212312286131;
     uint256 constant IC2y = 15505639799627325590655490925645629510237191912675051461164047905620055081525;
-    
+
     uint256 constant IC3x = 18884764497750409763397232717527940320640511300238447527286484703879507348436;
     uint256 constant IC3y = 4275035455714171418931465935819484853966010272184884395102075046670554594640;
-    
+
     uint256 constant IC4x = 20115639397728100122438186674280971695967279338290811699477377756213607556265;
     uint256 constant IC4y = 11706006804106781039332025887051799932356168495907189878250370997672263310654;
-    
+
     uint256 constant IC5x = 8521285627362779006350534057606830115741748579840881907434944662006658105510;
     uint256 constant IC5y = 16313318252176212977540975018977110902372399760035842580670661976409806961369;
-    
+
     uint256 constant IC6x = 11706855048712332607327616467099831449499495794846362372199880113724532319243;
     uint256 constant IC6y = 20020219776924857970475164979491676007123370130839575385339800711120874002914;
-    
+
     uint256 constant IC7x = 16376602012359952883643653589812472547408360535937694966694895314851078473714;
     uint256 constant IC7y = 14118366375712084062490501696638344767785344973339842261978811548549924123313;
-    
+
     uint256 constant IC8x = 6118592676552762206430605598975691301641755473528852479240855772397923656831;
     uint256 constant IC8y = 9820841858766352452430235218861538154014928897229232732863169095121400487589;
-    
+
     uint256 constant IC9x = 17752912369039372854361631386326649910613804560926758515111262467678018888969;
     uint256 constant IC9y = 2755201719231473752283915053135549870014185352630823116015338884068427676377;
-    
+
     uint256 constant IC10x = 3981706996933543294677775180304468944916585303490745691837724799879734114389;
     uint256 constant IC10y = 6603241891958835487443966686788013254970421780753160852509791560806572789620;
-    
+
     uint256 constant IC11x = 16307509728527432448358198371104255548226056811642397603921241612392927330548;
     uint256 constant IC11y = 16450460423348423120342419781943112070108109552302129163860355687279810427669;
-    
+
     uint256 constant IC12x = 313312698340002130249363204923066150292372446962159580664692949565755815417;
     uint256 constant IC12y = 17326074647723368630556201548004397519661601629257126643883004610686283245814;
-    
+
     uint256 constant IC13x = 5327138172955730221033013513896802356767192799609737439566341343240730183917;
     uint256 constant IC13y = 10173208076150374107006315590677637594954640552935605712299501414517229128962;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[13] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[13] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, r)) {
@@ -100,7 +103,7 @@ contract Groth16VerifierConsensus {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -135,33 +138,32 @@ contract Groth16VerifierConsensus {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_pVk, IC4x, IC4y, calldataload(add(pubSignals, 96)))
-                
+
                 g1_mulAccC(_pVk, IC5x, IC5y, calldataload(add(pubSignals, 128)))
-                
+
                 g1_mulAccC(_pVk, IC6x, IC6y, calldataload(add(pubSignals, 160)))
-                
+
                 g1_mulAccC(_pVk, IC7x, IC7y, calldataload(add(pubSignals, 192)))
-                
+
                 g1_mulAccC(_pVk, IC8x, IC8y, calldataload(add(pubSignals, 224)))
-                
+
                 g1_mulAccC(_pVk, IC9x, IC9y, calldataload(add(pubSignals, 256)))
-                
+
                 g1_mulAccC(_pVk, IC10x, IC10y, calldataload(add(pubSignals, 288)))
-                
+
                 g1_mulAccC(_pVk, IC11x, IC11y, calldataload(add(pubSignals, 320)))
-                
+
                 g1_mulAccC(_pVk, IC12x, IC12y, calldataload(add(pubSignals, 352)))
-                
+
                 g1_mulAccC(_pVk, IC13x, IC13y, calldataload(add(pubSignals, 384)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -187,7 +189,6 @@ contract Groth16VerifierConsensus {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -204,7 +205,6 @@ contract Groth16VerifierConsensus {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -214,39 +214,38 @@ contract Groth16VerifierConsensus {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
+
             checkField(calldataload(add(_pubSignals, 128)))
-            
+
             checkField(calldataload(add(_pubSignals, 160)))
-            
+
             checkField(calldataload(add(_pubSignals, 192)))
-            
+
             checkField(calldataload(add(_pubSignals, 224)))
-            
+
             checkField(calldataload(add(_pubSignals, 256)))
-            
+
             checkField(calldataload(add(_pubSignals, 288)))
-            
+
             checkField(calldataload(add(_pubSignals, 320)))
-            
+
             checkField(calldataload(add(_pubSignals, 352)))
-            
+
             checkField(calldataload(add(_pubSignals, 384)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
